address review: cleanup DOCA exporter
- Add [[nodiscard]] to getHostname()
- Use std::array for hostname buffer
- Move DocaSharedContext definition to .cpp, forward-declare in header
- Remove dead initialized_ member (constructor throws on failure)
- Make initializeDoca() throw directly instead of returning status
